backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.4
numpy==2.2.6
pandas==2.3.1
opencv-python==4.12.0.88
librosa==0.11.0
soundfile==0.13.1
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
scikit-learn==1.7.1
requests==2.32.4
pymongo==4.10.1
python-dotenv==1.1.1
Pillow==11.3.0
transformers==4.44.2
huggingface-hub==0.36.2
safetensors==0.7.0
kagglehub==0.3.12
